pi-server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
adafruit-circuitpython-neopixel==6.3.11
RPi.GPIO==0.7.1
